la till checkadmin på allt nästan, så funkt finns

+++ routes/web.php
@@ -26,25 +26,31 @@ use App\Http\Controllers\SubCommentController;
 |
 */
 // Books //
+Route::middleware(['auth', 'checkUserRoleAdmin'])->group(function ()
+{
+  Route::get('/books/create', [BookController::class, 'create']);
+  Route::post('/books', [BookController::class, 'store']);
+  Route::get('/books/{book}/edit', [BookController::class, 'edit']);
+  Route::put('/books/{book}', [BookController::class, 'update']);
+  Route::delete('/books/{book}', [BookController::class, 'destroy']);
+});
+
+
 Route::get('/books', [BookController::class, 'index']);
-Route::get('/books/create', [BookController::class, 'create'])->middleware('auth');
-Route::post('/books', [BookController::class, 'store'])->middleware('auth');
-Route::get('/books/{book}/edit', [BookController::class, 'edit'])->middleware('auth');
-Route::put('/books/{book}', [BookController::class, 'update'])->middleware('auth');
-Route::delete('/books/{book}', [BookController::class, 'destroy'])->middleware('auth');
+// Route::get('/books', [BookController::class, 'index'])->middleware('checkUserRoleAdmin');
 Route::get('/books/{book}', [BookController::class, 'show']);
 
 
 // Authors
 Route::get('/authors', [AuthorController::class, 'index']);                             //Show all authors
-Route::get('/authors/create', [AuthorController::class, 'create']);                     //Show create author form
-Route::post('/authors', [AuthorController::class, 'store']);                            //Store new author
-Route::get('/authors/{author}/edit', [AuthorController::class, 'edit']);                //Show edit author form
-Route::put('/authors/{author}', [AuthorController::class, 'update']);                   //Update author
-Route::get('/authors/{author}/delete', [AuthorController::class, 'confirm_delete']);    //Show delete author form
-Route::delete('/authors/{author}', [AuthorController::class, 'destroy']);               //Delete author
-Route::get('/authors/{author}/hide', [AuthorController::class, 'confirm_hide']);        //Show hide author form
-Route::put('/authors/{author}/hide', [AuthorController::class, 'hide']);                //Hide author
+Route::get('/authors/create', [AuthorController::class, 'create'])->middleware('checkUserRoleAdmin');                     //Show create author form
+Route::post('/authors', [AuthorController::class, 'store'])->middleware('checkUserRoleAdmin');                            //Store new author
+Route::get('/authors/{author}/edit', [AuthorController::class, 'edit'])->middleware('checkUserRoleAdmin');                //Show edit author form
+Route::put('/authors/{author}', [AuthorController::class, 'update'])->middleware('checkUserRoleAdmin');                   //Update author
+Route::get('/authors/{author}/delete', [AuthorController::class, 'confirm_delete'])->middleware('checkUserRoleAdmin');    //Show delete author form
+Route::delete('/authors/{author}', [AuthorController::class, 'destroy'])->middleware('checkUserRoleAdmin');               //Delete author
+Route::get('/authors/{author}/hide', [AuthorController::class, 'confirm_hide'])->middleware('checkUserRoleAdmin');        //Show hide author form
+Route::put('/authors/{author}/hide', [AuthorController::class, 'hide'])->middleware('checkUserRoleAdmin');                //Hide author
 Route::get('/authors/{author}', [AuthorController::class, 'show']);                     //Show selected author - keep as last
 
 // Users
@@ -52,16 +58,16 @@ Route::get('/users', [UserController::class, 'index']);
 
 
 //Genres
-Route::get('/genres', [GenreController::class, 'index']);                               //Show all genres
-Route::get('/genres/create', [GenreController::class, 'create']);                       //Show create genre form
-Route::post('/genres', [GenreController::class, 'store']);                              //Store new genre
-Route::get('/genres/{genre}/edit', [GenreController::class, 'edit']);                   //Show edit genre form
-Route::put('/genres/{genre}', [GenreController::class, 'update']);                      //Update genre
-Route::get('/genres/{genre}/delete', [GenreController::class, 'confirm_delete']);       //Show delete genre form
-Route::delete('/genres/{genre}', [GenreController::class, 'destroy']);                  //Delete genre
-Route::get('/genres/{genre}/hide', [GenreController::class, 'confirm_hide']);           //Show hide genre form
-Route::put('/genres/{genre}/hide', [GenreController::class, 'hide']);                   //Hide genre
-Route::get('/genres/{genre}', [GenreController::class, 'show']);                        //Show selected genre - keep as last
+Route::get('/genres', [GenreController::class, 'index'])->middleware('checkUserRoleAdmin');                               //Show all genres
+Route::get('/genres/create', [GenreController::class, 'create'])->middleware('checkUserRoleAdmin');                       //Show create genre form
+Route::post('/genres', [GenreController::class, 'store'])->middleware('checkUserRoleAdmin');                              //Store new genre
+Route::get('/genres/{genre}/edit', [GenreController::class, 'edit'])->middleware('checkUserRoleAdmin');                   //Show edit genre form
+Route::put('/genres/{genre}', [GenreController::class, 'update'])->middleware('checkUserRoleAdmin');                      //Update genre
+Route::get('/genres/{genre}/delete', [GenreController::class, 'confirm_delete'])->middleware('checkUserRoleAdmin');       //Show delete genre form
+Route::delete('/genres/{genre}', [GenreController::class, 'destroy'])->middleware('checkUserRoleAdmin');                  //Delete genre
+Route::get('/genres/{genre}/hide', [GenreController::class, 'confirm_hide'])->middleware('checkUserRoleAdmin');           //Show hide genre form
+Route::put('/genres/{genre}/hide', [GenreController::class, 'hide'])->middleware('checkUserRoleAdmin');                   //Hide genre
+Route::get('/genres/{genre}', [GenreController::class, 'show'])->middleware('checkUserRoleAdmin');                        //Show selected genre - keep as last
 
 
 //Reviews
@@ -97,7 +103,7 @@ Route::get('/subcomments/hide/{subcomment}', [SubCommentController::class, 'conf
 Route::put('/subcomments/hide/{subcomment}', [SubCommentController::class, 'hide']);
 
 //Register
-Route::get('/register', [RegistrationController::class, 'showRegistrationForm']);
+// Route::get('/register', [RegistrationController::class, 'showRegistrationForm']);
 
 
 //Login
@@ -107,6 +113,9 @@ Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 Route::post('users/authenticate', [UserController::class, 'authenticate']);
 
+//The "->name('login')" method is used to give the route a specific name, 
+//which can be used to generate URLs for that route in other parts of the application. In this case, the route is named "login".
+
 // Route::get('/login', [LoginController::class, 'showLoginForm'])->name('login');
 // Route::post('/login', [LoginController::class, 'login']);
 
@@ -114,11 +123,11 @@ Route::post('users/authenticate', [UserController::class, 'authenticate']);
 // Route::get('/logout', [LoginController::class, 'logout'])->name('logout');
 
 //admin
-Route::get('/admin', [AdminController::class, 'index']);
-Route::get('/admin/{user}/edit', [AdminController::class, 'edit']);
-Route::get('/admin/{user}', [AdminController::class, 'show']);
-Route::put('/admin/{user}', [AdminController::class, 'update']);
-Route::get('/admin/info', [AdminController::class, 'info']);
+Route::get('/admin', [AdminController::class, 'index'])->middleware('checkUserRoleAdmin');
+Route::get('/admin/{user}/edit', [AdminController::class, 'edit'])->middleware('checkUserRoleAdmin');
+Route::get('/admin/{user}', [AdminController::class, 'show'])->middleware('checkUserRoleAdmin');
+Route::put('/admin/{user}', [AdminController::class, 'update'])->middleware('checkUserRoleAdmin');
+Route::get('/admin/info', [AdminController::class, 'info'])->middleware('checkUserRoleAdmin');
 
 //Home
 Route::get('/', [HomeController::class, 'index']);
